+++ epserde/src/impls/string.rs
@@ -13,8 +13,7 @@
 //! `String` and fully deserialize it as `Box<str>`.
 //!
 //! Similarly to the case of [slices], the [`SerInner`] implementation for
-//! `&str` is a serialization-only convenience: data serialized from a `&str`
-//! must be deserialized as `String` or `Box<str>`.
+//! `&str` is a serialization-only convenience.
 //!
 //! We implement [`TypeHash`] for `str` so that it can be used in
 //! [`PhantomData`].