+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [0.13.0] - Unreleased
+## [0.13.0] - 2026-07-10
 
 ### New